% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codegen.R
\name{construct_pipeline}
\alias{construct_pipeline}
\title{Construct full pipeline code}
\usage{
construct_pipeline(raw_data_name, steps)
}
\arguments{
\item{raw_data_name}{Name of the starting dataframe}

\item{steps}{List of code strings from modules}
}
\description{
Construct full pipeline code
}
