% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codegen.R
\name{generate_mutate_code}
\alias{generate_mutate_code}
\title{Generate mutate code}
\usage{
generate_mutate_code(new_col, expression)
}
\arguments{
\item{new_col}{Name of the new column}

\item{expression}{Expression string}
}
\description{
Generate mutate code
}
