% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_codegen.R
\name{generate_select_code}
\alias{generate_select_code}
\title{Generate select code}
\usage{
generate_select_code(cols)
}
\arguments{
\item{cols}{Vector of column names}
}
\description{
Generate select code
}
